Update tracer thalweg & surface figure aspect ratios and panel sizes.
The figure aspect ratio was changed to 16:9 to better match wide screen
displays.
The size ratio of the figure panels was changed to 1.618, the golden
ratio, for aesthetics.
The thalweg length and surface plot height were increased to include a
bit more of Johnstone Strait.
The surface plot width was increased to include the Squamish River
estuary.