Handle mismatched ferry sensor data array lengths.
When the sensor data array lengths on a device (e.g. OXYSENSOR) are
different (not clear how/why that happens), xarray inserts NaNs into
the shorter arrays when the device dataset object is created.
It does that to align the data variables to a single time-base.
This change filters out those NaNs so that the qaqcflag mask can be
applied.
It is necessary because the qaqcflag attribute of the sensor data
doesn't get the same NaN insertion treatment as the sensor data.