Add in the routine to calculate the CHS near surface currents file
Currents are upper 5-grid points and upper 10-grid points, averaged,
unstaggered and rotate to East and North.

Subroutine calculates the currents and put a file in the same directory.

Works on nowcast (nowcast-blue), forecast, and forecast2.

Happens after download_results.  Presently it is in tandem with updating
the rolling forecast.  Needs to be moved after once the rolling forecast
handles these files.