Fix update_forecast_dataset re: wwatch3 foreacst2
Jamie Mclean (ECCC) confirmed my suspicion that the rolling forecast was not
being correctly updated after the wwatch3 forecast2 runs. Specifically, the 24
hours from the previous day's forecast that are required to make the time series
contiguous were missing. That issue is an artifact of the change a while back
to running wwatch3 in separate nowcast and forecast runs.