Add fvcom-forecast option to ping_erddap worker
This has been missing for some time.
Its effect is that the lastModified value in the dataset metadata is often days
earlier than the most recent files in the dataset.
It wend unnoticed because the absence of the option produces messages in the
manager-stderr log, but no Sentry alerts because the command-line parser
exeption occurs before logging is set up.