Drop make_averaged_dataset from hindcast automation
The 15-18 `reshapr extract` jobs that this generated per optimum hindcast run
seems to be too much for the dask cluster on salish, resulting in an
unacceptably high job failure rate of ~20%.

Tests of day-by-day processing (3 concurrent reshapr jobs) as will happen when
we add make_averaged_dataset to automation for nowcast-green.202111 showed
nearly 100% reliability. So, processing of hindcast will be handled outside of
automation.