Update collect_weather worker re: watchdog=2.1.7
Replace the time.sleep(1) in the watchdog observer loop that collects
the expected files with joining the observer thread with a 1-second
timeout.

We need to have a timeout on the observer thread so that the status
of the expected files set gets checked, otherwise the worker never
finishes because the main thread is blocked by the observer thread.

I'm not really sure why this change is necessary, but it fixes the
problem of the worker not collecting files after the change to the
HRDPS continental grid.