Chg daily_river_flows grid dir relative path hack
The relative path to the grid/ clone is different in Doug's test automation for
the full test suite compared to what it is for just the test_daily_river_flows
module. We're done with running the test for that module in isolation, so
change the relative path to the value for the full test suite automation because
it is how it will be used most often now.