Change crop_gribs event handling to on_closed
Replaced the 'on_moved' event with the 'on_closed' event because file system
events that occur when HRDPS files are moved from downloads directory to
atmospheric forcing tree are create, several modify events, and close. That is
presumably because the files are being moved from one files system to another,
so the move is actually a copy/delete.