Add single variable-hour file cropping to crop_gribs
Allow the crop_gribs worker to crop a specific forecast variable for a
specific forecast hour. This is primarily intended for recovery from
crop_gribs stalls when only 1 or a few files need to be cropped to finish
processing. The feature is added through the use of optional --var-hour and
--var arguments. Test cases for this new functionality are included.