Add retry & timeout for crop_gribs stalled observer
If crop_gribs instance has been running for 8 hours the file system observer
thread may have stalled, or there may be a problem with the HRDPS data
dissemination.

If >10 files remain in the set of files to be cropped, it is probably due to a
data dissemination problem. So, log a critical message that includes
the set of uncropped files so that we can send it to MSC support.

If <=10 files remain, it is possible that the watchdog observer thread just
missed 1 or more files. So, log an info message and retry cropping the set of
uncropped files. If that fails with FileNotFoundError, it is probably due to a
data dissemination problem. So, log a critical message that includes the set of
uncropped files so that we can send it to MSC support.