Optimize xarray.open_mfdataset in make_ww3* workers
Add `chunks` and `drop_variables` parameters into `xarray.open_mfdataset()`
in both workers and tests to optimize data loading and memory usage. This update
improves efficiency and performance by limiting loading of unnecessary variables
and enables parallel processing for large datasets. The hope is that these
changes will resolve (or at least reduce the frequency of) the issue of one or
both of the `make_ww3*` workers stalling during the `ds.to_netcdf()` step in
which dask is computing the task graph.