Update `make_ww3_*_file` workers and their place in the automation workflow (#213)
* Add mpi hosts file to wave forecasts configuration

Enables wwatch3 to be run on different VMs to those where NEMO is run.
This is a step towards running wwatch3 nowcast/forecast runs concurrently with
nowcast-green.

* Optimize xarray.open_mfdataset in make_ww3* workers

Add `chunks` and `drop_variables` parameters into `xarray.open_mfdataset()`
in both workers and tests to optimize data loading and memory usage. This update
improves efficiency and performance by limiting loading of unnecessary variables
and enables parallel processing for large datasets. The hope is that these
changes will resolve (or at least reduce the frequency of) the issue of one or
both of the `make_ww3*` workers stalling during the `ds.to_netcdf()` step in
which dask is computing the task graph.

* Change to run wwatch3 concurrent with nowcast-green

Always launch make_ww3_*_file workers after NEMO forecast finishes instead of
using a config flag to control whether they launch then or after nowcast-green
runs. Concurrency is possible because we have changed to running wwatch3 on
different VMs to NEMO.

* Drop uses of 'wave_forecast_after' missed in 3e332b29