Replace 'ptrc_T' with 'biol_T' in make_plots
The 'ptrc_T' dataset in the make_plots.py script has been replaced with
'biol_T'. This update affects the assignment of 'biol_T_hr', and the variable
names used within the 'image_loops' for loop and associated function arguments.
This alteration is due to a change in the v202111 NEMO results file names.