Drop "nowcast-dev" run from automation workflow and configuration (#229)
* Drop XIOS-2 & NEMO-3.6 build from skookum deployment

The deployment documentation has been updated to remove references to and
instructions on building XIOS-2 and NEMO-3.6 because, as of V21-11, we no longer
do the daily nowcast-dev run on salish. This keeps the docs accurate and
relevant to the current project setup.

* Drop runs directory setup from skookum deployment

The deployment documentation has been updated to remove runs/ directory setup on
skookum because, as of V21-11, we no longer do the daily nowcast-dev run on
salish. This keeps the docs accurate and relevant to the current project setup.

* Rename skookum deployment docs file

`skookum_salish.rst` -> `skookum.rst` because, as of V21-11, we no longer do
the daily nowcast-dev run on salish. So, there is no deployment setup to do on
salish. This keeps the docs accurate and relevant to the current project setup.

* Drop "nowcast-dev" run from automation

As of V21-11, we no longer do the daily nowcast-dev run on salish.

* Update deployment docs re: end of nowcast-dev runs

* Restore nowcast-dev results archive config item

Retained until make_plots worker is refactored to remove nowcast-dev. That
refactoring needs to come after the compare_venus_ctd fig module is updated
to handle the ONC API v3.

* Restore shared storge config item for salish-nowcast

Retained because the next_workers module checks the shared storage item for
all enabled hosts. So, it can't be removed for salish-nowcast until that whole
stanza can be removed.

* Fix tests re: acad43d7283487