Automate creation of day and month averaged biology, chemistry & physics datasets (#242)
* Automate creation of day averaged datasets

Add launch of make_averaged_dataset work for var_groups="biology", "chemistry",
"physics" for "nowcast-green" after completion of results download. Also added
test cases for the functionality.

* Remove make_averaged_dataset from hindcast log

make_averaged_dataset is included in nowcast automation now, so it is included
in the nowcast.log files.

* Expand make_averaged_dataset success/failure msgs

The success and failure message types for 'make_averaged_dataset' have been made
more detailed. These now include information on time interval and variable group,
enhancing the granularity and precision of the logs. This revision impacts both
how messages are registered and how they are handled in the
'make_averaged_dataset' function and corresponding tests.

* Improve make_averaged_dataset checklist element

Move run date from checklist dict key to be an item in the dict. This allows
next_worker functions to more easily access the run date that the averaged
dataset is from.

* Automate creation of month-averaged datasets

Add launch of make_averaged_dataset worker for month-averaged datasets for
var_groups="biology", "chemistry", "physics" for "nowcast-green" after
completion of day-averaged dataset creation on the last day of the month. Also
added test cases for the functionality.

* Remove "host_name" arg from make_averaged_dataset worker

The "host_name" argument in the "make_averaged_dataset" worker was removed as
it's unnecessary. Tests and calls to NextWorker were updated accordingly to
reflect this change.