Change from 'graham-dtn' to 'robot.graham'
This commit changes all instances of the string 'graham-dtn' to
'robot.graham'. This change is necessary to work with the introduction of MFA
for logins on graham. The 'robot.graham' is a new node specifically for
automation in concert with restricted ssh keys. The ssh key for 'robot.graham'
is also updated to 'SalishSeaCast_robot.graham_ed25519'.