Update dask worker settings in deployment docs
This commit modifies the instructions in the deployment docs for the persistent
dask cluster on `salish` that the `make_averaged_dataset` worker uses. It
changes the number of threads for each worker from 4 to 1, the memory limit from
automatic to 64G, and worker files will now be stored on the /tmp/ file system
instead of /dev/shm shared memory file system.

The change to 1 thread per worker is a consequence of reading
https://github.com/xCDAT/xcdat/issues/561#issuecomment-1969470260. Changing the
memory limit and worker file storage are a result of research and testing on
1mar24 (see worklog for details).