Documentation maintenance (#248)
* Fix typo in skookum deployment docs page title

* Update dask commands from hyphenated to sub-commands

dask.distributed 2022.10.0 deprecated the `dask-scheduler` and `dask-worker` CLI
commands in favour of `dask scheduler` and `dask worker`.

* Update dask worker settings in deployment docs

This commit modifies the instructions in the deployment docs for the persistent
dask cluster on `salish` that the `make_averaged_dataset` worker uses. It
changes the number of threads for each worker from 4 to 1, the memory limit from
automatic to 64G, and worker files will now be stored on the /tmp/ file system
instead of /dev/shm shared memory file system.

The change to 1 thread per worker is a consequence of reading
https://github.com/xCDAT/xcdat/issues/561#issuecomment-1969470260. Changing the
memory limit and worker file storage are a result of research and testing on
1mar24 (see worklog for details).

* Drop autodoc mocks for nemo_nowcast workers

The NEMO_Nowcast package is now installed in the readthedocs build environment.
So, autodoc mocks for its imports are no longer required. Removing those mocks
silences warnings about mocked objects and missing attributes for the
`clear_checklist` and `rotate_logs` worker docs.

* Add 'reshapr' to autodoc mocks list

This resolves an import error that prevents generation of docs for the
`make_averaged_dataset` worker on readthedocs.

* Replace inappropriate kbd directives in skookum docs

Changed all the kbd directives in the skookum deployment documentation into
inline code. This helps to ensure the correct semantic representation of
technical terms and commands.

re: issue #126

* Improve semantic markup of tmux in skookum docs

* Improve semantic markup of tmux in skookum docs

* Replace inappropriate kbd directives in docs

Changed all the kbd directives in the documentation into inline code. This helps
to ensure the correct semantic representation of technical terms and commands.

re: issue #126

* Remove unused imports in wave_height_period.py

Two unused Python library imports, pathlib and requests, were removed from
wave_height_period.py. This cleanup improves the code readability and efficiency.

* Update fig dev docs re: `black` for code formatting

The documentation has been updated to reflect the change in the automatic code
formatting tool used by the `salishsea_site` and `SalishSeaNowcast` packages.
Previously, we were using `yapf`, but switched to `black`.

* Add SSH keys and config section to skookum docs

The documentation now contains a guide for generating passphrase-less RSA and
ED25519 key pairs for the remote hosts that SalishSeaCast uses. It includes
commands to install the public keys and to edit the ssh configuration file on
'skookum'.

re: issue #244