Update ONC VENUS nodes model/obs comparison figure module (#254)
* Add Jupyter 'notebook' pkg to fig dev environment

The commit adds the Jupyter 'notebook' package to the environment for figures
development in environment-fig-dev.yaml. It used to be installed as a dependency
of the 'jupyterlab' package, but now must be installed separately.

* Add docstring to compare_venus_ctd fig module

A docstring was added to compare_venus_ctd.py. It includes an explanation of the
figure that the module created and the link to its testing notebook.

* Handle missing dev run fields in compare_venus_ctd

The code in compare_venus_ctd.py has been updated to gracefully handle cases
when the development run fields dataset (dev_grid_T_hr) is missing or not
provided. Previously, missing development grid data would have caused an
exception. Now, the code tests for this condition and sets other corresponding
variables to None as necessary, preventing the figure generation from crashing.
This elimination of unwanted exceptions makes the application more robust and
easier to maintain.

* Update compare_venus_ctd to to use ONC API v3

The data_tools.get_onc_data() call in the compare_venus_ctd figure module was
updated to use the ONC API v3. Parameters "getByStation", "station",
"deviceCategory", and "sensors" have been replaced with "getByLocation",
"locationCode", "deviceCategoryCode", and "sensorCategoryCodes", respectively.

* Update handling of nowcast-dev configuration

This commit adjusts the handling of 'nowcast-dev' configuration entry in
'nowcast.yaml' and 'make_plots.py'. If 'nowcast-dev' results archive path is set
to 'None', the code no longer tries to work with non-existent directories. This
change is reflected in update of related tests as well.

* Remove ONC Delta DDL node from make_plots worker

The "Compare_VENUS_Delta_DDL" figure calculation has been removed from the
make_plots worker. The ONC DDL node deployment ended in Mar-2024.