Improve WWatch3 wind and currents forcing file generation (#271)
* Add h5netcdf pkg dependency

Hoping to improve the reliability of the make_ww3_current_file and
make_ww3_wind_file workers by changing them to use h5netcdf for dataset reads.

The change results in version updates for multiple dependencies in
requirements .txt, and inclusion of the h5netcdf package across multiple
environment files and the pyproject.toml file.

* Update make_ww3_*_file workers main() function docstrings

Removed not informative "Set up and run the worker." line at the beginning.

re: issue #121

* Update wind file generation re: unneeded variables

Refactored the WWatch3 wind file generation to remove unnecessary variables and
improved related tests. The code now drops more unneeded variables to reduce
the memory load. The corresponding tests are also enhanced to accurately
represent these changes.

* Switch to h5netcdf engine for xarray dataset reads

The h5netcdf engine has been set as the engine for opening datasets in
'make_ww3_wind_file.py'. The intent is to avoid the netcdf4 package
thread-safety issues. The corresponding test cases in
'test_make_ww3_wind_file.py' have also been updated to reflect this change.

* Increase wind time_counter chunk size from 1 to 24

The chunk size for the time_counter variable in the make_ww3_wind_file worker
and corresponding tests has been increased. This change is anticipated to
improve efficiency by processing data in larger batches.

* Use dask processes for wwatch3 wind file generation

The wind file creation process for the WWatch3 model has been updated to use
processes rather than the default threads for the dask scheduler. Processes have
been found to be more reliable for dask operations on the types of workloads we
use in SalishSeaCast.

* Use netcdf4 to save dataset in make_ww3_wind_file

Explicitly use netcdf4 as the engine for dataset writing. This avoids
incompatibilities in the resulting file that arise if it is written using
h5netcdf.

* Update make_ww3_current_file re: unneeded variables

Refactored the WWatch3 current file generation to remove unnecessary variables
and improved related tests. The code now drops more unneeded variables to reduce
the memory load. The corresponding tests are also enhanced to accurately
represent these changes.

* Switch to h5netcdf engine for xarray dataset reads

The h5netcdf engine has been set as the engine for opening datasets in
'make_ww3_current_file.py'. The intent is to avoid the netcdf4 package
thread-safety issues. The corresponding test cases in
'test_make_ww3_current_file.py' have also been updated to reflect this change.

* Decrease current time_counter chunk size to 1

The chunk size for the time_counter variable in the make_current_wind_file
worker has decreased from 3 to 1. Testing showed that the smaller chunk size
resulted in slightly faster processing.

* Use dask processes for wwatch3 currents file generation

The currents file creation process for the WWatch3 model has been updated to use
processes rather than the default threads for the dask scheduler. Processes have
been found to be more reliable for dask operations on the types of workloads we
use in SalishSeaCast.

* Use netcdf4 to save dataset in make_current_wind_file

Explicitly use netcdf4 as the engine for dataset writing. This avoids
incompatibilities in the resulting file that arise if it is written using
h5netcdf.