Handle ValueError in get_onc_ctd salinity resampling (#278)
* Handle ValueError in get_onc_ctd salinity resampling

xarray version 2023.05.0 introduced a ValueError exception raised when a
DataArray being resampled is empty. We previously handled that condition by
catching an IndexError. Now we catch ValueError too.

The change to ValueError for an empty salinity DataArray didn't show up in
production until the 14jun24 update of the production environment brought xarray
version 2024.06.0 into use.

* Correct spelling errors in get_onc_ctd logging msgs

A couple of typos were found in the logging messages in the get_onc_ctd worker.
Specifically, the word "exclude" was mistyped as "exlude". This has now been
corrected.