Update datetime format in VFPA HADCP data handling
The datetime parsing format was adjusted in the VFPA HADCP data handling module.
This change will ensure the timestamp is correctly interpreted as
YYYY-MM-DD HH:MM as opposed to the previous format DD/MM/YYYY HH:MM. This change
was necessitated by a change in the datetime format in the observation .csv
files that was made between 17:00 and 18:00 UTC on 2023-07-02.