Add new subtopic to hrdps-continental-dd-weather.conf
Include an additional subtopic pattern in anticipation of 27-Nov-2024 change
in the directory structure on the dd.weather.gc.ca server to match that of the
hpfx.collab.science.gc.ca server. This should ensure that there is no
interruption when the directory structure is changed. After the change the
`model_hrdps.continental.2.5km.#` can be removed.