Change sarracenia to use hpfx.collab.science.gc.ca
dd.weather.gc.ca was unreliable after its directory structure was change to
match that of hpfx on 4-Dec-2024. So, we switched back to using hpfx and found
that there were none of the missing file messages and duplicated files issues
in HRDPS downloads that caused us to switch to dd.weather on 25-Apr-2023.