Add new subtopic pattern for datamart directory structure change (#305)
* Add new subtopic to hrdps-continental-dd-weather.conf

Include an additional subtopic pattern in anticipation of 4-Dec-2024 change
in the directory structure on the dd.weather.gc.ca server to match that of the
hpfx.collab.science.gc.ca server. This should ensure that there is no
interruption when the directory structure is changed. After the change the
`model_hrdps.continental.2.5km.#` can be removed.

* Add new subtopic to hydrometric-dd-weather.conf

Include an additional subtopic pattern re: 4-Dec-2024 change
in the directory structure on the dd.weather.gc.ca server to match that of the
hpfx.collab.science.gc.ca server. After the change the
`model_hrdps.continental.2.5km.#` can be removed.

* Drop outdated subtopics from dd-weather conf files

Drop subtopic patterns that were disabled by the 4-Dec-2024 change
in the directory structure on the dd.weather.gc.ca server to match that of the
hpfx.collab.science.gc.ca server.

* Change sarracenia to use hpfx.collab.science.gc.ca

dd.weather.gc.ca was unreliable after its directory structure was change to
match that of hpfx on 4-Dec-2024. So, we switched back to using hpfx and found
that there were none of the missing file messages and duplicated files issues
in HRDPS downloads that caused us to switch to dd.weather on 25-Apr-2023.