Update to Python 3.13 (#302)
* Add Python 3.13 to GHA pytest-with-coverage workflow

This lets us use the workflow to test whether all the packages that
SalishSeaNowcast depends on have been updated to support Python 3.13.

* Upgrade the minimum Python version to 3.13

Updated environment specifications, documentation, and workflows to require
Python 3.13 instead of 3.12. Adjusted related links and annotations to reflect
the new Python version. This ensures compatibility with the latest Python
features and fixes.