Update grib_to_netcdf to use `valid_time`
Replaced usages of `step` and `time` with `valid_time` to make our
timestamp processing compatible with non-hourly step handling that
was introduced in `cfgrib=0.9.12.0`. Adjusted preprocessing, data
array creation, and coordinate handling to align with this change.

This change removes the need for version pinning of `cfgrib=0.9.11.0`
that was introduced in PR#276. By doing that, issue #336 is resolved.