Refactor grib_to_netcdf time_counter calculation
Update `time_counter` calculation to use forecast step ranges. This mitigates
the fact that `cfgrib` no longer provides correct forecast time step values from
ECCC GRIB2 files. This is a rework of commit 406e7e55 that tried to resolve the
issue using the `valid_time` GRIB variable.

This change makes our timestamp processing compatible with non-hourly step
handling that was introduced in `cfgrib=0.9.12.0`.

This change removes the need for version pinning of `cfgrib=0.9.11.0`
that was introduced in PR#276. By doing that, issue #336 is resolved.