Remove FVCOM support from update_forecast_datasets
This commit eliminates FVCOM-related configurations, logic, and tests from the
`update_forecast_datasets` worker. The FVCOM model is no longer a supported
option for rolling forecast updates because we are no longer running the VHFR
FVCOM model. This cleanup eliminates unused functionality and simplifies the
codebase.

Several instances of `unitest.mock.patch.dict()` in the test suite were replaced
with `pytest.monkeypatch` to simplify and modernize the tests.