Remove unused VHFR-FVCOM figure notebooks & modules
FVCOM results figures are no longer needed because we are no longer running the
VHFR-FVCOM model. The figure modules, their development and testing notebooks,
the code to generate the figures in the `make_plots` worker, and the associated
configuration elements have been removed. This cleanup eliminates unused
functionality and simplifies the codebase.