Remove references to OPPTools and its dependencies
This commit eliminates all references to the no longer used OPPTools package
across documentation, configuration, and dependencies. Additionally, it removes
related dependencies such as `pyproj`, `utm`, and `zeep` for a cleaner and more
maintainable environment setup. The OPPTools package was a dependency for the
FVCOM run preparation workers that is no longer needed because we are no longer
running the VHFR-FVCOM model.