Update collect_river_data checklist structure re: issue #170 (#354)
* Update collect_river_data checklist structure

Modified the checklist to replace "river name" item with the river name from the
config and the day-averaged flow file path using `os.fspath`. This addresses
issue #170 by making the river name key in the checklist unique for every river.
That ensures that the river names and the files in which their data were
collected all appear in the checklist instead of just the last river that was
processed.

* Add SPDX license identifier to collect_river_data

Added the SPDX-License-Identifier: Apache-2.0 to comply with licensing
requirements. This ensures better license clarity and automatic license detection
for the file. No functional changes were made to the code.

* Refactor test functions to improve readability

Add consistent spacing between logical blocks in test functions for better
readability. This minor refactor does not alter test functionality but ensures
the code is easier to maintain and understand.

* Refactor discharge calculation test fixture mocks

Refactored `mock_calc_eccc_day_avg_discharge` and
`mock_get_usgs_day_avg_discharge` into staticmethod pytest fixtures to improve
reusability and clarity. Also updated the `test_checklist` method to use the new
fixtures, simplifying the test setup.

* Add test for collect_river_data log messages

A new test `test_log_messages` was added to verify log output during the river
data collection process. This ensures the logging behaviour is consistent and
captures the expected debug and info messages.