Fix wave height and period obs processing (#360)
Figure generation was failing when there were missing observations in the data
obtained from NDBC. It appears that NDBC recently changed the missing data
indicator to "MM". This commit handles missing values in wave height and dominant
period data by filtering out "MM" entries and casting the Series to correct data
types. The casting is necessary because the "MM" causes the Series to have
"object" dtypes instead of numeric ones.