Fix indentation of test assertions
Corrected the indentation for assertions in
`test_run_NEMO_hindcast .test_no_job_found_on_queue()`` to
move them into the `pytest.raises()` block scope. This ensures the
assertions are applied within the intended context. It is hoped that this
will resolved the issue of the test failing in the GHA workflow even though
it succeeds when run locally.