Fix indentation of test assertions
Corrected the indentation for assertions in more `test_run_NEMO_hindcast`
test functions to move them into the `pytest.raises()` block scope. This
ensures the assertions are applied within the intended context. It is hope
that this will resolve the issue of the test failing in the GHA workflow
even though it succeeds when run locally.