Update ERDDAP weather dataset ID in config & tests
Revised the ERDDAP `weather` dataset ID from
`ubcSSaSurfaceAtmosphereFieldsV1` to
`ubcSSaSurfaceAtmosphereFieldsV23-02` in `nowcast.yaml` and the
corresponding test file. This was missed when we switched to the
HRDPS continental rotated lon-lat product, but the oversight was
hidden by ERDDAP doing dataset updates every 10 minutes. The ERDDAP
behaviour was recently dropped, resulting in this issue being
surfaced.