Updates to work on new `orcinus` HPC cluster configuration (#386)
Merged not because all the work to get `nowcast-agrif` running on the new `orcinus` configuration is complete, but because it has stalled. Rebuilding the executable is going to take a lot of work because the module system has largely gone away on `orcinus` and its unclear how I get the libraries I need. The priority of getting the 201702 configuration running again is very low.

* Update orcinus forcing dir paths in config & tests

Replaced `/home/sallen/MEOPAR/` paths with `/global/home/sallen/MEOPAR/` for
`orcinus` forcing directories in `nowcast.yaml` and associated test files. This
maintains consistency with the updated directory structure on the `orcinus`
cluster.

* Add missing SPDX short-form license identifier

Inserted the `SPDX-License-Identifier: Apache-2.0` comment at the top of the
`make_forcing_links.py` file. This communicates license information in portable
and machine-readable manner.

* Update orcinus run prep dir paths in config & tests

Updated `run prep dir` paths for `orcinus-nowcast-agrif` in `nowcast.yaml` and
the corresponding test file. This ensures alignment with the updated directory
structure on the `orcinus` cluster.

* Add unit tests for production YAML config elements

Introduced tests for production config elements used by the `run_NEMO_agrif`
worker. These ensure production YAML configuration consistency and correctness
for the worker functionality.

re: issue #117.

* Update `salishsea` command path in config & tests

Revised `salishsea cmd` paths for `orcinus-nowcast-agrif` in `nowcast.yaml` and
its associated test file. This change ensures consistency with the updated
directory structure on the `orcinus` cluster.