Update host name from `robot.graham` to `robot.nibi` (#403)
* Update host name from `robot.graham` to `robot.nibi`

Replaced all occurrences of `robot.graham` with `robot.nibi` in configuration
and test files. This change reflects the replacement of the `graham` HPC
cluster with the `nibi` cluster.

* Update ssh key file names for `robot.nibi`

Only the names of the ssh keys have been changed, not the keys themselves,
because Alliance staff migrated the configuration for our key from
`robot .graham` to `robot.nibi`. This change reflects the replacement of the
`graham` HPC cluster with the `nibi` cluster.